Code cleanup 3- 2025
